+++ packages/fluekit/src/Icons.ts
@@ -49,4 +49,5 @@ export const Icons = {
   monitor:
     "M20 3H4c-1.1 0-2 .9-2 2v11c0 1.1.9 2 2 2h3l-1 1v2h12v-2l-1-1h3c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2zm0 13H4V5h16v11z",
   warning: "M1 21h22L12 2 1 21zm12-3h-2v-2h2v2zm0-4h-2v-4h2v4z",
+  videoPlay: "M8 5v14l11-7z",
 };